+++ src/components/Page/Mac/mac.scss
@@ -11,9 +11,24 @@
     margin: 1em 0;
 
     &-title {
-      font-size: 1.5em;
-      font-weight: 600;
-      margin-bottom: .5em;
+      width: 100%;
+      margin: 1em 0;
+
+      & h2 {
+        font-size: 1.5em;
+        font-weight: 600;
+        //margin-bottom: .5em;
+        width: 100%; 
+        text-align: center; 
+        border-bottom: 1px solid rgba(0, 0, 0, 0.4); 
+        line-height: 0.1em;
+        margin: 10px 0 20px; 
+
+        & span {
+          background:#fff; 
+          padding:0 10px; 
+        }
+      }
     }
 
     &-content {
