+++ src/components/Page/index.js
@@ -7,13 +7,9 @@ import './page.scss';
 import Cards from '../../containers/Page/Cards';
 import Mac from './Mac';
 
-import { getThreeMostAffected, getPourcentageEvolution, getNameOfActiveCountry } from '../../utils/functions';
+import { getThreeMostAffected, getPourcentageEvolution } from '../../utils/functions';
 
-const Page = ({ countries, activeCountry, statsData, globalStats, dailyStats, fetchStatsData, onSetActiveCountry, onLoadStatsCountry, onLoadEvolutionStats }) => {
-  // Set countries options for the Select a Country Component
-  const [countriesOptions, setCountriesOptions] = useState([]);
-  // Set Active country to show data only for this country
-  const [activeCountryName, setActiveCountryName] = useState('Global');
+const Page = ({ countries, countriesOptions, activeCountry, statsData, globalStats, dailyStats, fetchStatsData, onLoadEvolutionStats }) => {
   // Set Most Affected Countries stats 
   const [mostAffectedCountries, setmostAffectedCountries] = useState([]);
   // Set Active Range for Select a Range Componenet
@@ -22,25 +18,6 @@ const Page = ({ countries, activeCountry, statsData, globalStats, dailyStats, fe
   const [activeEvolution, setActiveEvolution] = useState();
 
   useEffect(() => {
-    let countriesArray = [{
-      key: 'GLO',
-      value: 'GLO',
-      text: 'Global'
-    }];
-
-    if (countries !== null) {
-      countries.map((country) => {
-        const data = {
-          key: country.iso3,
-          value: country.iso3,
-          text: country.name,
-        }
-        countriesArray.push(data);
-      })
-    }
-
-    setCountriesOptions(countriesArray);
-
     if (globalStats != null & countries !== null ) {
       let mac =(getThreeMostAffected(globalStats, countriesOptions));
       setmostAffectedCountries(mac);
@@ -54,35 +31,10 @@ const Page = ({ countries, activeCountry, statsData, globalStats, dailyStats, fe
     setActiveEvolution(getPourcentageEvolution(globalStats, dailyStats, countries, activeCountry))
     
   }, [activeRange, dailyStats, activeCountry])
-
-  const loadStatsCountry = (data) => {
-    // Set the redux state
-    onSetActiveCountry({
-      iso: data,
-      name: getNameOfActiveCountry(countriesOptions, data),
-    })
-
-    if (data === 'GLO') {
-      fetchStatsData();
-    } else {
-      onLoadStatsCountry(data);
-    }
-  }
   
   return (
     <div id="page">
       <div className="page">
-        
-        <div className="page-select">
-          <Select 
-            placeholder='Select a country' 
-            options={countriesOptions} 
-            defaultValue={activeCountry.iso}
-            onChange={(e, data) => {
-              loadStatsCountry(data.value)
-            }}
-          />
-        </div>
 
         { statsData != undefined && 
           (
